refactor by separation of logic
