+++ agenta-cli/agenta/cli/variant_commands.py
@@ -1,7 +1,6 @@
 import os
 import re
 import sys
-from typing import List
 from pathlib import Path
 
 from requests.exceptions import ConnectionError
@@ -9,9 +8,9 @@ from requests.exceptions import ConnectionError
 import click
 import questionary
 import toml
-from agenta.cli import helper
+from agenta.cli import helper, variant_configs, command_utils
 from agenta.cli.telemetry import event_track
-from agenta.client.api_models import AppVariant, Image
+from agenta.client.api_models import Image
 from agenta.docker.docker_utils import build_tar_docker_container
 
 from agenta.client.api import add_variant_to_server
@@ -24,30 +23,6 @@ BACKEND_URL_SUFFIX = os.environ.get("BACKEND_URL_SUFFIX", "api")
 def variant():
     """Commands for variants"""
     pass
-
-
-def read_config_file(file_path):
-    try:
-        with open(file_path, "r") as file:
-            config_data = toml.load(file)
-        return config_data
-    except Exception as e:
-        print(f"Error reading the config file: {e}")
-        return None
-
-def extract_parameters(config_data):
-    parameters = {}
-    if "parameters" in config_data:
-        parameters = config_data["parameters"]
-    else:
-        click.echo(
-            click.style(
-                f"Parameters not found in config file. Please make sure you have a parameters section in the config file.",
-                fg="red",
-            )
-        )
-        
-    return parameters
 
 
 def add_variant(
@@ -173,13 +148,17 @@ def add_variant(
                         fg="bright_black",
                     )
                 )
-                variant_id = config["variant_ids"][config["variants"].index(variant_name)]
+                variant_id = config["variant_ids"][
+                    config["variants"].index(variant_name)
+                ]
                 client.update_variant_image(
                     variant_id=variant_id,
                     request=image,  # because Fern code uses "request: Image" instead of "image: Image"
                 )  # this automatically restarts
             else:
-                click.echo(click.style(f"Adding {variant_name} to server...", fg="yellow"))
+                click.echo(
+                    click.style(f"Adding {variant_name} to server...", fg="yellow")
+                )
                 response = add_variant_to_server(
                     app_id, base_name, image, f"{host}/{BACKEND_URL_SUFFIX}", api_key
                 )
@@ -243,40 +222,37 @@ def add_variant(
             )
         # Last step us to save the config file
         toml.dump(config, config_file.open("w"))
-        if overwrite:
-            # In the case we are overwriting, don't return anything. Otherwise the command server would attempt to start the container which would result in an error!!!
-            # TODO: Improve this stupid design
-            return None
-        else:
-            return variant_id
     else:
         # get app base id from backend
         client = AgentaApi(
             base_url=f"{host}/{BACKEND_URL_SUFFIX}",
             api_key=api_key,
         )
-            
+
         get_base_id = client.list_bases(app_id=app_id)
-        
+
         # Check if the list is not empty
         if get_base_id:
             first_base = get_base_id[0]
             base_id = first_base.base_id
         else:
             click.echo(click.style("No bases found.", fg="red"))
-        
+
         # get parameters from config file
         variant_config_file = app_path / file_name
-        config_data = read_config_file(variant_config_file)
-        parameters = extract_parameters(config_data)
         
-        payload = {
-            "base_id": base_id,
-            "new_variant_name": config_name,
-            "new_config_name": config_name.split(".")[1],
-            "parameters": parameters,
-        }
+        try:
+            config_data = helper.read_config_file(variant_config_file)
+        except Exception as ex:
+            click.echo(click.style(f"{ex}", fg="red"))
+            return
         
+        try:
+            parameters = helper.extract_parameters(config_data)
+        except Exception as ex:
+            click.echo(click.style(f"{ex}", fg="red"))
+            return
+
         # create a new variant
         try:
             click.echo(
@@ -285,181 +261,46 @@ def add_variant(
                     fg="bright_black",
                 )
             )
-            response = client.add_variant_from_base_and_config(payload=payload)
+            response = client.add_variant_from_base_and_config(
+                base_id=base_id,
+                new_variant_name=config_name,
+                new_config_name=config_name,
+                parameters=parameters,
+            )
+            variant_name = "app" + "." + config_name
+            variant_id = response.variant_id
+            click.echo(
+                click.style(
+                    f"New variant {config_name} created successfully.",
+                    fg="green",
+                )
+            )
         except Exception as ex:
             click.echo(click.style(f"Error while creating variant: {ex}", fg="red"))
+            return
+
+    # pull config for the new variant
+    try:
+        command_utils.pull_config_from_backend(
+            config=config,
+            app_folder=app_folder,
+            api_key=api_key,
+            variant_names=[variant_name],
+            host=host,
+            new_config=True,
+            variant_id=variant_id,
+        )
+    except Exception as ex:
+        click.echo(click.style(f"Error while pulling config: {ex}", fg="red"))
+        return
+
+    if config_name == "default":
+        if overwrite:
+            # In the case we are overwriting, don't return anything. Otherwise the command server would attempt to start the container which would result in an error!!!
+            # TODO: Improve this stupid design
             return None
-
-
-def start_variant(variant_id: str, app_folder: str, host: str):
-    """
-    Starts a container for an existing variant
-    Args:
-        variant_name: the name of the variant
-        app_folder: the folder of the app
-    """
-    app_folder = Path(app_folder)
-    config_file = app_folder / "config.toml"
-    config = toml.load(config_file)
-    app_id = config["app_id"]
-    api_key = config.get("api_key", "")
-
-    if len(config["variants"]) == 0:
-        click.echo("No variants found. Please add a variant first.")
-        return
-
-    if variant_id:
-        if variant_id not in config["variant_ids"]:
-            click.echo(
-                click.style(
-                    f"Variant {variant_id} not found in backend. Maybe you removed it in the webUI?",
-                    fg="red",
-                )
-            )
-            return
-    else:
-        variant_name = questionary.select(
-            "Please choose a variant", choices=config["variants"]
-        ).ask()
-        variant_id = config["variant_ids"][config["variants"].index(variant_name)]
-
-    client = AgentaApi(
-        base_url=f"{host}/{BACKEND_URL_SUFFIX}",
-        api_key=api_key,
-    )
-
-    endpoint = client.start_variant(variant_id=variant_id, action={"action": "START"})
-    click.echo("\n" + click.style("Congratulations! 🎉", bold=True, fg="green"))
-    click.echo(
-        click.style("Your app has been deployed locally as an API. 🚀", fg="cyan")
-        + click.style(" You can access it here: ", fg="white")
-        + click.style(f"{endpoint}/", bold=True, fg="yellow")
-    )
-
-    click.echo(
-        click.style("\nRead the API documentation. 📚", fg="cyan")
-        + click.style(" It's available at: ", fg="white")
-        + click.style(f"{endpoint}/docs", bold=True, fg="yellow")
-    )
-
-    webui_host = "http://localhost" if host == "localhost" else host
-    click.echo(
-        click.style(
-            "\nStart experimenting with your app in the playground. 🎮",
-            fg="cyan",
-        )
-        + click.style(" Go to: ", fg="white")
-        + click.style(f"{webui_host}/apps/{app_id}/playground", bold=True, fg="yellow")
-        + "\n"
-    )
-
-
-def remove_variant(variant_name: str, app_folder: str, host: str):
-    """
-    Removes a variant from the server
-    Args:
-        variant_name: the name of the variant
-        app_folder: the folder of the app
-    """
-    config_file = Path(app_folder) / "config.toml"
-    config = toml.load(config_file)
-    app_name = config["app_name"]
-    api_key = config.get("api_key", "")
-
-    if not config["variants"]:
-        click.echo(
-            click.style(
-                f"No variants found for app {app_name}. Make sure you have deployed at least one variant.",
-                fg="red",
-            )
-        )
-        return
-
-    if variant_name:
-        if variant_name not in config["variants"]:
-            click.echo(
-                click.style(
-                    f"Variant {variant_name} not found in backend. Maybe you already removed it in the webUI?",
-                    fg="red",
-                )
-            )
-            return
-    else:
-        variant_name = questionary.select(
-            "Please choose a variant", choices=config["variants"]
-        ).ask()
-        if not variant_name:
-            click.echo("Operation cancelled.")
-            sys.exit(0)
-    variant_id = config["variant_ids"][config["variants"].index(variant_name)]
-
-    variant_configuration_file = Path(app_folder) / f"{variant_name}.toml"
-    if not variant_configuration_file.exists():
-        pass
-
-    client = AgentaApi(
-        base_url=f"{host}/{BACKEND_URL_SUFFIX}",
-        api_key=api_key,
-    )
-
-    try:
-        client.remove_variant(variant_id=variant_id)
-
-        # delete the variant configuration file if it exists
-        if variant_configuration_file.is_file():
-            click.echo(
-                click.style(
-                    f"Removing variant configuration file {variant_name}.toml...",
-                    fg="bright_black",
-                )
-            )
-            variant_configuration_file.unlink()
-    except Exception as ex:
-        click.echo(
-            click.style(
-                f"Error while removing variant {variant_name} for App {app_name} from the backend",
-                fg="red",
-            )
-        )
-        click.echo(click.style(f"Error message: {ex}", fg="red"))
-        return
-
-    click.echo(
-        click.style(
-            f"Variant {variant_name} for App {app_name} removed successfully from Agenta!",
-            fg="green",
-        )
-    )
-
-
-def list_variants(app_folder: str, host: str):
-    """List available variants for an app and print them to the console
-
-    Arguments:
-        app_folder -- _description_
-    """
-    config_file = Path(app_folder) / "config.toml"
-    config = toml.load(config_file)
-    app_name = config["app_name"]
-    app_id = config["app_id"]
-    api_key = config.get("api_key", "")
-    variants = []
-
-    client = AgentaApi(
-        base_url=f"{host}/{BACKEND_URL_SUFFIX}",
-        api_key=api_key,
-    )
-
-    try:
-        variants: List[AppVariant] = client.list_app_variants(app_id=app_id)
-    except Exception as ex:
-        raise ex
-
-    if variants:
-        for variant in variants:
-            helper.display_app_variant(variant)
-    else:
-        click.echo(click.style(f"No variants found for app {app_name}", fg="red"))
+        else:
+            return variant_id
 
 
 def config_check(app_folder: str):
@@ -480,20 +321,8 @@ def config_check(app_folder: str):
             )
         )
         return
-    host = get_host(app_folder)  # TODO: Refactor the whole config thing
-    helper.update_config_from_backend(config_file, host=host)
-
-
-def get_host(app_folder: str) -> str:
-    """Fetches the host from the config"""
-    app_folder = Path(app_folder)
-    config_file = app_folder / "config.toml"
-    config = toml.load(config_file)
-    if "backend_host" not in config:
-        host = "http://localhost"
-    else:
-        host = config["backend_host"]
-    return host
+    host = helper.get_host(app_folder)  # TODO: Refactor the whole config thing
+    helper.update_config_from_backend(config_file, host=host, app_folder=app_folder)
 
 
 @variant.command(name="remove")
@@ -504,10 +333,10 @@ def remove_variant_cli(variant_name: str, app_folder: str):
 
     try:
         config_check(app_folder)
-        remove_variant(
+        command_utils.remove_variant(
             variant_name=variant_name,
             app_folder=app_folder,
-            host=get_host(app_folder),
+            host=helper.get_host(app_folder),
         )
     except Exception as ex:
         click.echo(click.style(f"Error while removing variant: {ex}", fg="red"))
@@ -545,7 +374,7 @@ def serve_cli(ctx, app_folder: str, file_name: str):
         return
 
     try:
-        host = get_host(app_folder)
+        host = helper.get_host(app_folder)
     except Exception as e:
         click.echo(click.style("Failed to retrieve the host.", fg="red"))
         click.echo(click.style(f"Error message: {str(e)}", fg="red"))
@@ -567,7 +396,9 @@ def serve_cli(ctx, app_folder: str, file_name: str):
 
     if variant_id:
         try:
-            start_variant(variant_id=variant_id, app_folder=app_folder, host=host)
+            command_utils.start_variant(
+                variant_id=variant_id, app_folder=app_folder, host=host
+            )
         except ConnectionError:
             error_msg = "Failed to connect to Agenta backend. Here's how you can solve the issue:\n"
             error_msg += "- First, please ensure that the backend service is running and accessible.\n"
@@ -586,7 +417,16 @@ def list_variants_cli(app_folder: str):
     """List the variants in the backend"""
     try:
         config_check(app_folder)
-        list_variants(app_folder=app_folder, host=get_host(app_folder))
+        variants, app_name = command_utils.list_variants(
+            app_folder=app_folder, host=helper.get_host(app_folder)
+        )
+
+        if variants:
+            for variant in variants:
+                helper.display_app_variant(variant)
+        else:
+            click.echo(click.style(f"No variants found for app {app_name}", fg="red"))
+
     except Exception as ex:
         click.echo(click.style(f"Error while listing variants: {ex}", fg="red"))
 
@@ -600,13 +440,16 @@ def list_variants_cli(app_folder: str):
 )
 @click.option("--app_folder", default=".")
 @click.option("--config_file", help="The name of the config to use")
-@click.option("--from_variant", help="The name of the variant to base the new variant on")
+@click.option(
+    "--from_variant", help="The name of the variant to base the new variant on"
+)
 @click.pass_context
 def add_variant_from_config(ctx, app_folder: str, config_file: str, from_variant: str):
     """Adds a variant to the backend"""
     config_commands = ctx.args
     variant_to_use = None
     variant_config_file = None
+    new_varaint_name = None
 
     # check and update config file
     try:
@@ -623,7 +466,6 @@ def add_variant_from_config(ctx, app_folder: str, config_file: str, from_variant
     error_msg += ">>> agenta variant add --from_variant app.<variant_name> === To add a variant based on an existing variant.\n"
     error_msg += "or\n"
     error_msg += ">>> agenta variant add --config app.<variant_name>.toml === To add a variant based on an existing config file.\n"
-    
 
     # be double sure that there's an app and an existing variant in the config file
     app_folder = Path(app_folder)
@@ -645,27 +487,25 @@ def add_variant_from_config(ctx, app_folder: str, config_file: str, from_variant
         )
         sys.exit(0)
 
-
     # validate that only one of the options is provided
     if not config_commands:
         if not from_variant and not config_file:
-            
             variant_to_use = questionary.select(
-                "Please choose a variant to use as a base for the new variant", choices=app_config.get("variants")
+                "Please choose a variant to use as a base for the new variant",
+                choices=app_config.get("variants"),
             ).ask()
             if not variant_to_use:
                 click.echo("Operation cancelled.")
                 sys.exit(0)
+
         elif from_variant and not config_file:
-            
             variant_to_use = from_variant
-            
+
         elif config_file and not from_variant:
-            
             if len(config_file.split(" ")) > 1:
                 click.echo(click.style(f"{error_msg}", fg="red"))
                 sys.exit(0)
-            
+
             # validate that name of the config_file is in the format app.<variant_name>.toml
             if not re.match("^[a-zA-Z0-9_]+\.[a-zA-Z0-9_]+\.toml$", config_file):
                 click.echo(
@@ -675,16 +515,28 @@ def add_variant_from_config(ctx, app_folder: str, config_file: str, from_variant
                     )
                 )
                 sys.exit(0)
-            
-            variant_config_file = Path(app_folder) / config_file
-            if not variant_config_file.exists():
+
+            # validate that a previous variant with the same name does not exist
+            if config_file.removesuffix(".toml") in app_config.get("variants"):
                 click.echo(
                     click.style(
-                        f"No config file by name {config_file} exists! Please make sure you are in the right directory and that you have created the config file in the right directory.", 
+                        f"Variant {config_file.removesuffix('.toml')} already exists. Please choose another name.",
                         fg="red",
                     )
                 )
                 sys.exit(0)
+
+            variant_config_file = Path(app_folder) / config_file
+            if not variant_config_file.exists():
+                click.echo(
+                    click.style(
+                        f"No config file by name {config_file} exists! Please make sure you are in the right directory and that you have created the config file in the right directory.",
+                        fg="red",
+                    )
+                )
+                sys.exit(0)
+            new_varaint_name = config_file.split(".")[1]
+
         elif from_variant and config_file:
             click.echo(click.style(f"{error_msg}", fg="red"))
             sys.exit(0)
@@ -692,18 +544,14 @@ def add_variant_from_config(ctx, app_folder: str, config_file: str, from_variant
         if len(config_commands) > 0:
             click.echo(click.style(f"{error_msg}", fg="red"))
             sys.exit(0)
-            
-        if config_commands and ((from_variant and config_file) or from_variant or config_file):
+
+        if config_commands and (
+            (from_variant and config_file) or from_variant or config_file
+        ):
             click.echo(click.style(f"{error_msg}", fg="red"))
             sys.exit(0)
-    
-        
+
     if variant_to_use is not None:
-        click.echo(
-            click.style(
-                f"Adding a new variant to the backend\n variant_to_use in condition: {variant_to_use}", fg="bright_black"
-            )
-        )
         # get and validate variant name
         if len(variant_to_use.split(".")) != 2:
             click.echo(
@@ -718,32 +566,72 @@ def add_variant_from_config(ctx, app_folder: str, config_file: str, from_variant
         if variant_to_use not in app_config.get("variants"):
             click.echo(
                 click.style(
-                    f"Variant {variant_to_use} not found in backend. Maybe you already removed it in the webUI?",
+                    f"Base variant to use: {variant_to_use} not found in backend. Maybe you already removed it in the webUI?",
                     fg="red",
                 )
             )
             return
-        
+
         # validate that variant config file exists
         from_variant_config_file = app_folder / f"{variant_to_use}.toml"
         if not from_variant_config_file.exists():
             click.echo(
                 click.style(
-                    f"Config file for variant {variant_to_use} not found. Please first run 'agenta config pull {from_variant}' to pull the config file from the backend.",
+                    f"Config file for variant {variant_to_use} not found. Please first run 'agenta config pull {variant_to_use}' to pull the config file from the backend.",
                     fg="red",
                 )
             )
             return
-        
         variant_config_file = from_variant_config_file
+
+    if not config_file:
+        new_varaint_name = questionary.text(
+            "Please enter a name for the new variant"
+        ).ask()
+        if not new_varaint_name:
+            click.echo("Operation cancelled.")
+            sys.exit(0)
+
+    def validate_variant_name(variant_name):
+        # validate regex that variant name contains only alphanumeric characters, underscores, and hyphens
+        if not re.match("^[a-zA-Z0-9_-]+$", variant_name):
+            click.echo(
+                click.style(
+                    "Invalid input. Please use only alphanumeric characters, underscores, and hyphens in the variant name.",
+                    fg="red",
+                )
+            )
+            return False
+
+        # validate that variant name does not exist
+        existing_variants = app_config.get("variants")
+        for existing_variant in existing_variants:
+            if existing_variant.split(".")[1] == variant_name:
+                click.echo(
+                    click.style(
+                        f"Variant {new_varaint_name} already exists. Please choose another name.",
+                        fg="red",
+                    )
+                )
+                return False
+
+        return True
+
+    if new_varaint_name is not None:
+        while not validate_variant_name(new_varaint_name):
+            new_varaint_name = questionary.text(
+                "Please enter a name for the new variant"
+            ).ask()
+            if not new_varaint_name:
+                click.echo("Operation cancelled.")
+                sys.exit(0)
 
     try:
         add_variant(
             app_folder=app_folder,
             file_name=variant_config_file,
-            host=get_host(app_folder),
-            config_name=variant_config_file.name.removesuffix(".toml"),
+            host=helper.get_host(app_folder),
+            config_name=new_varaint_name,
         )
     except Exception as ex:
         click.echo(click.style(f"Error while adding variant: {ex}", fg="red"))
-
